+++ packages/common/src/types.ts
@@ -1,5 +1,5 @@
 import { type StrategyOptions } from '@octokit/auth-app';
-import type { github_teams, repocop_vulnerabilities } from '@prisma/client';
+import type { repocop_vulnerabilities } from '@prisma/client';
 
 export type GithubAppSecret = {
 	appId: string;
@@ -72,11 +72,3 @@ export const SLAs: Record<Severity, number | undefined> = {
 };
 
 export type NonEmptyArray<T> = [T, ...T[]];
-
-type TeamFields = Pick<github_teams, 'slug' | 'id' | 'name'>;
-
-export interface Team extends TeamFields {
-	slug: NonNullable<TeamFields['slug']>;
-	id: NonNullable<TeamFields['id']>;
-	name: NonNullable<TeamFields['name']>;
-}
